Refactor diff report builder * Violation Reporting are now in own file * more helper methods in base class HtmlReportBuilder
