+++ lib/pmdtester/builders/diff_report_builder.rb
@@ -7,14 +7,9 @@ module PmdTester
   # Building diff report for a single project
   class DiffReportBuilder < HtmlReportBuilder
     include PmdTester
-    include DiffReportBuilderConfigErrors
+    include DiffReportBuilderViolations
     include DiffReportBuilderErrors
-
-    def initialize
-      @a_index = 1
-      @b_index = 1
-      @c_index = 1
-    end
+    include DiffReportBuilderConfigErrors
 
     def build(project)
       @project = project
@@ -54,15 +49,7 @@ module PmdTester
 
     def build_summary_table(doc)
       doc.table(class: 'bodyTable', border: '0') do
-        doc.thead do
-          doc.tr do
-            doc.th 'Item'
-            doc.th 'Base'
-            doc.th 'Patch'
-            doc.th 'Difference'
-          end
-        end
-
+        build_table_head(doc, 'Item', 'Base', 'Patch', 'Difference')
         build_summary_table_body(doc)
       end
     end
@@ -107,85 +94,6 @@ module PmdTester
           doc.text @project.get_path_inside_project(filename)
         end
       end
-    end
-
-    def build_violations_section(doc, violation_diffs)
-      doc.div(class: 'section', id: 'Violations') do
-        doc.h2 'Violations:'
-
-        doc.h3 HtmlReportBuilder::NO_DIFFERENCES_MESSAGE if violation_diffs.empty?
-        violation_diffs.each do |key, value|
-          doc.div(class: 'section') do
-            build_filename_h3(doc, key)
-            build_violation_table(doc, key, value)
-          end
-        end
-      end
-    end
-
-    def build_violation_table(doc, key, value)
-      doc.table(class: 'bodyTable', border: '0') do
-        build_violation_table_head(doc)
-        build_violation_table_body(doc, key, value)
-      end
-    end
-
-    def build_violation_table_head(doc)
-      doc.thead do
-        doc.tr do
-          doc.th
-          doc.th 'priority'
-          doc.th 'Rule'
-          doc.th 'Message'
-          doc.th 'Line'
-        end
-      end
-    end
-
-    def build_violation_table_body(doc, key, value)
-      doc.tbody do
-        value.each do |pmd_violation|
-          build_violation_table_row(doc, key, pmd_violation)
-        end
-      end
-    end
-
-    def build_violation_table_row(doc, key, pmd_violation)
-      doc.tr(class: pmd_violation.branch == BASE ? 'b' : 'a') do
-        # The anchor
-        doc.td do
-          doc.a(id: "A#{@a_index}", href: "#A#{@a_index}") { doc.text '#' }
-          @a_index += 1
-        end
-
-        violation = pmd_violation.attrs
-
-        # The priority of the rule
-        doc.td violation['priority']
-
-        # The rule that trigger the violation
-        doc.td do
-          doc.a(href: (violation['externalInfoUrl']).to_s) { doc.text violation['rule'] }
-        end
-
-        # The violation message
-        doc.td pmd_violation.text
-
-        # The begin line of the violation
-        line = violation['beginline']
-
-        # The link to the source file
-        doc.td do
-          link = get_link_to_source(violation, key)
-          doc.a(href: link.to_s) { doc.text line }
-        end
-      end
-    end
-
-    def get_link_to_source(violation, key)
-      l_str = @project.type == 'git' ? 'L' : 'l'
-      line_str = "##{l_str}#{violation['beginline']}"
-      @project.get_webview_url(key) + line_str
     end
   end
 end
