+++ test/resources/html_report_builder/expected_diff_report_index.html
@@ -61,7 +61,7 @@
 <table class="bodyTable" border="0">
 <thead><tr>
 <th></th>
-<th>priority</th>
+<th>Priority</th>
 <th>Rule</th>
 <th>Message</th>
 <th>Line</th>
@@ -82,7 +82,7 @@ Fields should be declared at the top of the class, before any method declaration
 <table class="bodyTable" border="0">
 <thead><tr>
 <th></th>
-<th>priority</th>
+<th>Priority</th>
 <th>Rule</th>
 <th>Message</th>
 <th>Line</th>
@@ -123,7 +123,7 @@ Fields should be declared at the top of the class, before any method declaration
 <table class="bodyTable" border="0">
 <thead><tr>
 <th></th>
-<th>priority</th>
+<th>Priority</th>
 <th>Rule</th>
 <th>Message</th>
 <th>Line</th>
@@ -144,7 +144,7 @@ Fields should be declared at the top of the class, before any method declaration
 <table class="bodyTable" border="0">
 <thead><tr>
 <th></th>
-<th>priority</th>
+<th>Priority</th>
 <th>Rule</th>
 <th>Message</th>
 <th>Line</th>
@@ -165,7 +165,7 @@ Fields should be declared at the top of the class, before any method declaration
 <table class="bodyTable" border="0">
 <thead><tr>
 <th></th>
-<th>priority</th>
+<th>Priority</th>
 <th>Rule</th>
 <th>Message</th>
 <th>Line</th>
